iot-health-backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
motor==3.3.2
pymongo==4.6.1
pydantic==2.6.4
python-dotenv==1.0.1
scikit-learn
numpy
